MuseScore_Generator/requirements.txt
music21==6.8.1
pillow==9.3.0